+++ bin/buildall.sh
@@ -5,5 +5,5 @@ set -e
 
 lerna run --stream --scope=@bitaccess/coinlib-common build
 lerna run --stream --scope=@bitaccess/coinlib-bitcoin build
-lerna run --stream --scope=@bitaccess/coinlib-* --ignore=@bitaccess/coinlib-payments --ignore=@bitaccess/coinlib-bitcoin --parallel build
+lerna run --stream --scope=@bitaccess/coinlib-* --ignore=@bitaccess/coinlib-payments --ignore=@bitaccess/coinlib-bitcoin --ignore=@bitaccess/coinlib-common --parallel build
 lerna run --stream --scope=@bitaccess/coinlib-payments build
