+++ packages/coinlib-ethereum/package.json
@@ -74,7 +74,7 @@
     "web3-core": "^1.3.6"
   },
   "dependencies": {
-    "@bitaccess/coinlib-payments": "file:../coinlib-payments",
+    "@bitaccess/coinlib-common": "file:../coinlib-common",
     "@faast/ts-common": "^0.6.0",
     "bignumber.js": "^9.0.0",
     "bip32": "^2.0.5",
