+++ packages/coinlib-litecoin/package.json
@@ -76,7 +76,7 @@
   },
   "dependencies": {
     "@bitaccess/coinlib-bitcoin": "file:../coinlib-bitcoin",
-    "@bitaccess/coinlib-payments": "file:../coinlib-common",
+    "@bitaccess/coinlib-common": "file:../coinlib-common",
     "@faast/ts-common": "^0.6.0",
     "@types/lodash": "^4.14.150",
     "bignumber.js": "^9.0.0",
